add a windows_executable
